Move validate option to plugin
Also add _HOOMD_SKIP_CPU_TESTS_WHEN_GPUS_PRESENT_ env var to replace the old TEST_CPU_IN_GPU_BUILDS option.
